requirements.txt
pyjwt
fire
pymongo
tqdm